Extract Device.hpp into focused modules
Split the 1071-line Device.hpp into 8 files by responsibility:
- NetworkConfig.hpp: MQTT broker + NTP + instance/location config
- HeapTrace.hpp: conditional heap tracing/task-tracking debug utilities
- MqttCommands.hpp: basic, NVS, and HTTP update MQTT command registration
- ConfigUpdate.hpp: config reconciliation (applyConfigUpdate, registerUpdateHandler)
- SyncPublisher.hpp: SYNC message publishing and trigger task
- TelemetryTask.hpp: periodic telemetry collection and publishing
- BootHelpers.hpp: factory reset, battery/NVS/watchdog/MQTT init helpers
- Device.hpp: startDevice<T>() boot sequence orchestration (519 lines, down from 1071)

Pass firmwareVersion explicitly to registerUpdateHandler() and initSyncTask()
instead of relying on the file-level static variable.
